fix: Refactor exceptions
