+++ lib/sio_core.dart
@@ -1,3 +1,4 @@
 export 'src/mnemonic.dart';
 export 'src/build_transaction.dart';
 export 'src/broadcast.dart';
+export 'src/exceptions.dart';
